split ray.hpp
