+++ src/ray.h
@@ -2,12 +2,10 @@
 #define RAY_H
 
 #include <vector>
-#include <cmath>
-#include <algorithm>
 #include <memory>
 #include <optional>
+#include <algorithm>
 #include "matrix.hpp"
-#include "transformations.h"
 #include "shapes.hpp"
 
 namespace RT {
@@ -33,40 +31,11 @@ struct Ray {
     Matrix::Vector4f direction{};
 };
 
-Matrix::Vector4f Position(const Ray& ray, const float t) {
-    return ray.origin + ray.direction * t;
-}
+Matrix::Vector4f Position(const Ray& ray, const float t);
 
-Ray Transform(const Ray& ray, const Matrix::Matrix4f& transformation) {
-    return Ray{transformation * ray.origin, transformation * ray.direction};
-}
+Ray Transform(const Ray& ray, const Matrix::Matrix4f& transformation);
 
-std::vector<Intersection<Shapes::Sphere>> Intersect(const Shapes::Sphere& sphere, const Ray& ray) {
-    Ray ray2 = Transform(ray, sphere.GetTransform().Inverse());
-    Matrix::Vector4f sphereToRay = ray2.origin - Point(0, 0, 0);
-
-    float a = Dot(ray2.direction, ray2.direction);
-    float b = 2 * Dot(ray2.direction, sphereToRay);
-    float c = Dot(sphereToRay, sphereToRay) - 1;
-    float discriminant = b*b - 4*a*c;
-
-    std::vector<Intersection<Shapes::Sphere>> xs{};
-    if (discriminant < 0)
-        // No intersections
-        return xs;
-
-    // One or two intersections
-    float t1 = (-b - std::sqrt(discriminant)) / (2 * a);
-    float t2 = (-b + std::sqrt(discriminant)) / (2 * a);
-
-    if (t1 > t2)
-        std::swap(t1, t2);
-
-    xs.push_back(Intersection<Shapes::Sphere>{t1, sphere});
-    xs.push_back(Intersection<Shapes::Sphere>{t2, sphere});
-
-    return xs;    
-}
+std::vector<Intersection<Shapes::Sphere>> Intersect(const Shapes::Sphere& sphere, const Ray& ray);
 
 template<typename Shape>
 std::vector<Intersection<Shape>> Intersections(std::same_as<Intersection<Shape>> auto& ...inter) {
@@ -84,7 +53,7 @@ std::optional<Intersection<Shape>> Hit(const std::vector<Intersection<Shape>>& i
     for (const Intersection<Shape>& i : intersections) {
         if (i.t < 0)
             continue;
-        return std::make_optional(i);   // Return smalleste nonnegative value
+        return std::make_optional(i);   // Return smallest nonnegative value
     }
     return std::nullopt;
 }
